requirements.txt
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.119.0
google-api-core==2.26.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.11.1
lxml==6.0.2
multidict==6.7.0
openai==2.5.0
packaging==25.0
postgrest==2.22.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
PyJWT==2.10.1
pyparsing==3.2.5
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.1.1
realtime==2.22.0
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
storage3==2.22.0
StrEnum==0.4.15
supabase==2.22.0
supabase-auth==2.22.0
supabase-functions==2.22.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
yarl==1.22.0
